requirements/base.txt
# Bleeding edge Django
django==1.11.8

# Configuration
django-environ==0.4.4

# Forms
django-braces==1.9.0
django-crispy-forms==1.7.0
django-floppyforms==1.7.0

# Models
django-model-utils==3.0.0

# Images
Pillow==4.3.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.34.0

# Time zones support
pytz==2016.6.1

# Redis support
django-redis==4.8.0

celery==3.1.23

# Webpack
django-webpack-loader==0.5.0

# Your custom requirements go here
